+++ mre/src/Materials/MPathTracerMat.h
@@ -9,23 +9,23 @@ class MPathTracerMat : public Material {
 public:
 	MPathTracerMat();
 	MPathTracerMat(Vec3 c);
-	MPathTracerMat(Vec3 c, Vec3 e, double i, int t);
-	MPathTracerMat(Vec3 c, Vec3 e, double i, int t, double g);
+	MPathTracerMat(Vec3 c, Vec3 e, float i, int t);
+	MPathTracerMat(Vec3 c, Vec3 e, float i, int t, float g);
 
 	Vec3 getColor();
 	Vec3 getEmission();
-	double getIOR();
+	float getIOR();
 	int getType();
-	double getGamma();
+	float getGamma();
 
 	std::string toString();
 
 private:
 	Vec3 color;
 	Vec3 emissionColor;
-	double IOR;
+	float IOR;
 	int type; //0 diffuse, 1 pure specular, 2 glass
-	double gamma;
+	float gamma;
 
 
 };
